pyproject.toml
[project]
name = "honeycrisp"
version = "0.1.0"
description = "Cider Fermenting Pipeline"
authors = [
]
dependencies = [
    "transformers>=4.41.2",
    "torch>=2.3.1",
    "typer>=0.12.3",
    "pandas>=2.2.2",
    "mlflow-skinny>=2.14.0",
    "pyarrow>=16.1.0",
    "sentence-transformers>=3.0.1",
    "matplotlib>=3.9.0",
    "tqdm>=4.66.4",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.isort]
profile = "black"
known_first_party = []

[tool.docformatter]
recursive = true
black = true

[tool.pytest.ini_options]
addopts = "--cov=honeycrisp"
testpaths = ["tests"]

[tool.ruff]
lint.select = [
    "E", # Pycodestyle
    "ERA", # Eradicate
    "F", # Pyflakes
    "T20", # Print
    "TRY002", # Create own exception
    "TRY004", # Type check without TypeError
    "TRY201", # verbose raise
    "TRY301", # raise with try
    "TRY302", # useless try except
    "TRY400", # logging.error used over logging.exception
    "TRY401", # verbose log message
    "Q", # Quotes
    "DTZ", # Naive datetime
]

[tool.pdm.dev-dependencies]
dev = [
    "jupyter==1.0.0",
    "pytest>=8.2.2",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.setuptools]
packages = ["honeycrisp"]
